// github/types/paths/orgs/{org}/memberships/{username}.types.ts
export type orgsGetMembershipForUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: orgsGetMembershipForUser_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: org_membership;
          };
        };
        examples: {
          "response-if-user-has-an-active-admin-membership-with-organization": unknown;
        };
      };
      403: forbidden;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type orgsSetMembershipForUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: orgsSetMembershipForUser_Path;
    headers: never;
    cookie: never;
    body: {
      /**
       * The role to give the user in the organization. Can be one of:
       *  * `admin` - The user will become an owner of the organization.
       *  * `member` - The user will become a non-owner member of the organization.
       * @default "member"
       */
      role?: "admin" | "member";
    };
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: org_membership;
          };
        };
        examples: {
          "response-if-user-already-had-membership-with-organization": unknown;
        };
      };
      403: forbidden;
      422: validation_failed;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type orgsRemoveMembershipForUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: orgsRemoveMembershipForUser_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      403: forbidden;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type orgsGetMembershipForUser_Path = { org: string; username: string };

export type orgsSetMembershipForUser_Path = { org: string; username: string };

export type orgsRemoveMembershipForUser_Path = {
  org: string;
  username: string;
};